protocols/kad: Make automatic kbucket update on connect optional
Add an option to `KademliaConfig` to prevent Kademlia from automatically
adding a node to its kbuckets whenever it is notified of a new
connection to that node.

This enables a user of Kademlia to control which nodes are part of the
local view of the DHT. E.g. in the case of Substrate this enables one to
exclude light-clients from the local view of the DHT.

The default behaviour stays untouched.